+++ cli/integration-test/integration/diagnostics_test.clj
@@ -7,6 +7,52 @@
 
 (lsp/clean-after-test)
 
+(deftest different-aliases
+  (lsp/start-process!)
+  (lsp/request! (fixture/initialize-request))
+  (lsp/notify! (fixture/initialized-notification))
+  (lsp/notify! (fixture/did-open-source-path-notification "diagnostics/different_aliases.clj"))
+
+  (testing "When there are different aliases for the same namespace"
+    (h/assert-submaps
+      [{:range
+        {:start {:line 3, :character 34}, :end {:line 3, :character 35}},
+        :tags [],
+        :message
+        "Different aliases #{s string cstring str} found for clojure.string",
+        :code "clojure-lsp/different-aliases",
+        :langs [],
+        :severity 1,
+        :source "clj-kondo"}
+       {:range
+        {:start {:line 4, :character 34}, :end {:line 4, :character 37}},
+        :tags [],
+        :message
+        "Different aliases #{s string cstring str} found for clojure.string",
+        :code "clojure-lsp/different-aliases",
+        :langs [],
+        :severity 1,
+        :source "clj-kondo"}
+       {:range
+        {:start {:line 5, :character 34}, :end {:line 5, :character 40}},
+        :tags [],
+        :message
+        "Different aliases #{s string cstring str} found for clojure.string",
+        :code "clojure-lsp/different-aliases",
+        :langs [],
+        :severity 1,
+        :source "clj-kondo"}
+       {:range
+        {:start {:line 6, :character 34}, :end {:line 6, :character 41}},
+        :tags [],
+        :message
+        "Different aliases #{s string cstring str} found for clojure.string",
+        :code "clojure-lsp/different-aliases",
+        :langs [],
+        :severity 1,
+        :source "clj-kondo"}]
+      (lsp/client-awaits-server-diagnostics "diagnostics/different_aliases.clj"))))
+
 (deftest unused-public-var
   (lsp/start-process!)
   (lsp/request! (fixture/initialize-request))